[Refactor] 탐색 캐시 evict annotation 정리
