+++ api/src/main/kotlin/com/nexters/api/service/NewsletterExploreService.kt
@@ -1,5 +1,6 @@
 package com.nexters.api.service
 
+import com.nexters.api.util.LocalCache
 import com.nexters.external.repository.ExploreContentRow
 import com.nexters.external.service.ExposureContentService
 import org.springframework.stereotype.Service
@@ -35,14 +36,19 @@ class NewsletterExploreService(
         }
 
     private fun findCachedFirstExploreContentPage(size: Int): ExploreContentPageResult =
-        ExploreContentsCache.getFirstPage(size) {
+        LocalCache.getOrPut(
+            key = buildExploreContentPageCacheKey(size),
+            ttl = EXPOSURE_CONTENTS_CACHE_TTL_MINUTES,
+        ) {
             loadExploreContentPage(FIRST_PAGE_OFFSET, size)
         }
 
     private fun countExposureContents(): Long =
-        ExploreContentsCache.getTotalCount {
-            exposureContentService.countAllExposureContents()
-        }
+        LocalCache.getOrPut(
+            key = TOTAL_COUNT_CACHE_KEY,
+            ttl = EXPOSURE_CONTENTS_CACHE_TTL_MINUTES,
+            loader = exposureContentService::countAllExposureContents,
+        )
 
     private fun loadExploreContentPage(
         lastSeenOffset: Long,
@@ -92,5 +98,12 @@ class NewsletterExploreService(
     companion object {
         private const val MIN_PAGE_SIZE = 1
         private const val FIRST_PAGE_OFFSET = 0L
+        private const val EXPOSURE_CONTENTS_CACHE_TTL_MINUTES = 6 * 60L
+        private const val CACHE_KEY_PREFIX = "exposure:contents"
+        private const val TOTAL_COUNT_CACHE_KEY = "$CACHE_KEY_PREFIX:total-count"
+        private const val PAGE_CACHE_KEY_PREFIX = "$CACHE_KEY_PREFIX:page:"
+
+        private fun buildExploreContentPageCacheKey(size: Int): String =
+            "${PAGE_CACHE_KEY_PREFIX}last-seen-offset:$FIRST_PAGE_OFFSET:size:$size"
     }
 }
