+++ api/src/test/kotlin/com/nexters/api/controller/NewsletterApiControllerTest.kt
@@ -2,9 +2,9 @@ package com.nexters.api.controller
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.nexters.api.dto.CreateContentApiRequest
-import com.nexters.api.service.ExploreContentsCache
 import com.nexters.api.service.NewsletterContentsService
 import com.nexters.api.service.NewsletterExploreService
+import com.nexters.api.util.LocalCache
 import com.nexters.api.util.TokenUtil
 import com.nexters.external.entity.AdminMember
 import com.nexters.external.entity.Content
@@ -485,6 +485,10 @@ class NewsletterApiControllerTest {
         )
 
     private fun deleteExploreCacheKeys() {
-        ExploreContentsCache.evict()
+        LocalCache.deleteByPrefix(EXPLORE_CACHE_KEY_PREFIX)
+    }
+
+    private companion object {
+        private const val EXPLORE_CACHE_KEY_PREFIX = "exposure:contents:"
     }
 }
