+++ external/src/main/kotlin/com/nexters/external/annotation/ExposureContentChanged.kt
@@ -1,0 +1,5 @@
+package com.nexters.external.annotation
+
+@Target(AnnotationTarget.FUNCTION)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ExposureContentChanged
